requirements.txt
dateUts==0.2.7
iniUts==1.1.3
loguru==0.7.2
openpyxl==3.1.3
pandas==2.2.2
psutil==5.9.6
psycopg2==2.9.6
pycron==3.0.0
python-dotenv==1.0.1
selenium==4.15.2
schedule==1.1.0
undetected-chromedriver==3.5.5
urllib3==2.2.1
vaultUts==1.0.1